+++ .changeset/rotten-berries-press.md
@@ -1,0 +1,5 @@
+---
+"@gillsdk/react": minor
+---
+
+update to esm over commonJS, including updating relative imports
