+++ packages/react/src/__typeset__/latest-blockhash.ts
@@ -1,5 +1,5 @@
 import { GetLatestBlockhashApi } from "gill";
-import { useLatestBlockhash } from "../hooks";
+import { useLatestBlockhash } from "../hooks/index.js";
 
 // [DESCRIBE] useLatestBlockhash
 {
