+++ packages/react/src/__typeset__/signatures-for-address.ts
@@ -1,5 +1,5 @@
 import { Address, GetSignaturesForAddressApi } from "gill";
-import { useSignaturesForAddress } from "../hooks";
+import { useSignaturesForAddress } from "../hooks/index.js";
 
 // [DESCRIBE] useSignaturesForAddress
 {
