+++ packages/react/src/hooks/account.ts
@@ -3,9 +3,9 @@
 import { useQuery } from "@tanstack/react-query";
 import type { Account, Address, Decoder, FetchAccountConfig, Simplify } from "gill";
 import { assertAccountExists, decodeAccount, fetchEncodedAccount } from "gill";
-import { GILL_HOOK_CLIENT_KEY } from "../const";
-import { useSolanaClient } from "./client";
-import type { GillUseRpcHook } from "./types";
+import { GILL_HOOK_CLIENT_KEY } from "../const.js";
+import { useSolanaClient } from "./client.js";
+import type { GillUseRpcHook } from "./types.js";
 
 type RpcConfig = Simplify<Omit<FetchAccountConfig, "abortSignal">>;
