+++ packages/react/src/hooks/index.ts
@@ -1,12 +1,12 @@
-export * from "./account";
-export * from "./balance";
-export * from "./client";
-export * from "./latest-blockhash";
-export * from "./program-accounts";
-export * from "./recent-prioritization-fees";
-export * from "./signature-statuses";
-export * from "./signatures-for-address";
-export * from "./slot";
-export * from "./token-account";
-export * from "./token-mint";
-export * from "./transaction";
+export * from "./account.js";
+export * from "./balance.js";
+export * from "./client.js";
+export * from "./latest-blockhash.js";
+export * from "./program-accounts.js";
+export * from "./recent-prioritization-fees.js";
+export * from "./signature-statuses.js";
+export * from "./signatures-for-address.js";
+export * from "./slot.js";
+export * from "./token-account.js";
+export * from "./token-mint.js";
+export * from "./transaction.js";
