+++ packages/react/src/hooks/signatures-for-address.ts
@@ -2,9 +2,9 @@
 
 import { useQuery } from "@tanstack/react-query";
 import { Address, GetSignaturesForAddressApi, Simplify } from "gill";
-import { GILL_HOOK_CLIENT_KEY } from "../const";
-import { useSolanaClient } from "./client";
-import { GillUseRpcHook } from "./types";
+import { GILL_HOOK_CLIENT_KEY } from "../const.js";
+import { useSolanaClient } from "./client.js";
+import { GillUseRpcHook } from "./types.js";
 
 type RpcConfig = Simplify<Parameters<GetSignaturesForAddressApi["getSignaturesForAddress"]>[1]>;
