+++ packages/react/src/hooks/token-account.ts
@@ -1,9 +1,9 @@
 "use client";
 
 import { useQuery } from "@tanstack/react-query";
-import { GILL_HOOK_CLIENT_KEY } from "../const";
-import { useSolanaClient } from "./client";
-import type { GillUseRpcHook } from "./types";
+import { GILL_HOOK_CLIENT_KEY } from "../const.js";
+import { useSolanaClient } from "./client.js";
+import type { GillUseRpcHook } from "./types.js";
 
 import type { Account, Address, FetchAccountConfig, Simplify } from "gill";
 import { address, assertAccountExists, assertIsAddress, fetchEncodedAccount } from "gill";
