+++ packages/react/src/index.ts
@@ -1,4 +1,5 @@
-// @ts-ignore - Patch BigInt to allow calling `JSON.stringify` on objects that use
+// @ts-expect-error - Patch BigInt to allow calling `JSON.stringify` on objects that use
+// eslint-disable-next-line no-unused-vars
 interface BigInt {
   /** Convert a BigInt to string form when calling `JSON.stringify()` */
   toJSON: () => string;
@@ -12,6 +13,6 @@ if (BigInt.prototype.toJSON === undefined) {
   };
 }
 
-export * from "./const";
-export * from "./providers";
-export * from "./hooks";
+export * from "./const.js";
+export * from "./hooks/index.js";
+export * from "./providers.js";
